requirements.txt
flask
anthropic
pdfplumber
Pillow
python-docx
pdf2image
gunicorn
